requirements.txt
findspark
kafka-python
py4j
pyspark
spark
